refactor summation classes
